fixing expected output syntax

+++ programs/all-testers/M4/testers/decls/multi/floattype.go
@@ -1,8 +1,8 @@
-//~ +1.000000e+000+2.000000e+000
+//~+1.000000e+000+2.000000e+000
 
 package main
 
 func main() {
-	var a, b float = 1., 2.
+	var a, b float64 = 1., 2.
 	print(a,b)
 }
