+++ programs/all-testers/M4/testers/scoping/block/sub.go
@@ -1,4 +1,4 @@
-//~ 0
+//~0
 
 package main
 
@@ -6,7 +6,7 @@ func main() {
 
 	var a = "0"
 	{
-		var a float = 0.
+		var a float64 = 0.
 		a = a - 1;
 	}
 	print(a);
