+++ programs/all-testers/M4/testers/scoping/func/useglobal.go
@@ -1,10 +1,10 @@
-//~ +0.000000e+000
+//~+0.000000e+000
 
 package main
 
 var a = 0.;
 
-func somefunc(b float) float {
+func somefunc(b float64) float {
 	return b + a + 1.
 }
